+++ Sources/Socket.swift
@@ -37,6 +37,34 @@ public class Socket {
     /// is set to true
     public var autoReconnect: Bool = true
     
+    /// Enable/Disable SSL certificate validation by setting the value on the 
+    /// underlying WebSocket.
+    /// See https://github.com/daltoniam/Starscream#self-signed-ssl
+    public var disableSSLCertValidation: Bool {
+        get { return connection.disableSSLCertValidation }
+        set { connection.disableSSLCertValidation = newValue }
+    }
+
+    #if os(Linux)
+    #else
+    /// Configure custom SSL validation logic, eg. SSL pinning, by setting the 
+    /// value on the underlyting WebSocket.
+    /// See https://github.com/daltoniam/Starscream#ssl-pinning
+    public var security: SSLTrustValidator? {
+        get { return connection.security }
+        set { connection.security = newValue }
+    }
+    
+    /// Configure the encryption used by your client by setting the allowed 
+    /// cipher suites supported by your server.
+    /// See https://github.com/daltoniam/Starscream#ssl-cipher-suites
+    public var enabledSSLCipherSuites: [SSLCipherSuite]? {
+        get { return connection.enabledSSLCipherSuites }
+        set { connection.enabledSSLCipherSuites = newValue }
+    }
+    #endif
+    
+    
     
     //----------------------------------------------------------------------
     // MARK: - Private Attributes